refactoring criteria field and operator files
