--- apps/modernization-ui/src/design-system/input/operator/index.ts
@@ -1,1 +1,0 @@
-export { OperatorInput } from './OperatorInput';
