+++ apps/modernization-ui/src/design-system/input/text/criteria/criteria.module.scss
@@ -1,4 +1,4 @@
-.operatorInput {
+.criteriaField {
     display: flex;
     flex-direction: row;
     gap: 0.5rem;
