+++ apps/modernization-ui/src/design-system/input/text/criteria/index.ts
@@ -1,0 +1,1 @@
+export { TextCriteriaField } from './TextCriteriaField';
