+++ apps/modernization-ui/src/design-system/input/text/criteria/operator/index.ts
