+++ apps/modernization-ui/src/design-system/select/index.ts
@@ -1,3 +1,2 @@
 export * from './single/SingleSelect';
 export * from './multi/MultiSelect';
-export * from './operator/OperatorSelect';
